+++ lua/neotest-haskell/stack.lua
@@ -1,23 +1,24 @@
 local hspec = require('neotest-haskell.hspec')
--- local logger = require("neotest.logging")
+local logger = require("neotest.logging")
 
 local M = {}
 
+---@async
 function M.build_command(project_root, package_root, pos)
-  error("neotest-haskell: Stack not supported yet.") -- TODO
-  -- logger.debug('Building spec for Stack project...')
-  -- local is_subpackage = package_root ~= project_root
-  -- local package_name = vim.fn.fnamemodify(package_root, ':t')
-  -- local package_arg = is_subpackage and { package_name } or {}
-  -- local command = vim.tbl_flatten({
-  --   'stack',
-  --   'test',
-  --   package_arg,
-  --   '--ta',
-  --   '"--match \\"' .. hspec_match .. '\\"', -- FIXME
-  -- })
-  -- vim.notify('(async) Running: stack test ' .. package_arg .. ' --ta ' .. hspec_match)
-  -- return command
+  logger.debug('Building spec for Stack project...')
+  local command = {
+    'stack',
+    'test',
+  }
+  local is_subpackage = package_root ~= project_root
+  local package_name = vim.fn.fnamemodify(package_root, ':t')
+  if is_subpackage then
+    table.insert(command, package_name)
+  end
+  local test_opts = hspec.get_stack_test_opts(pos)
+  return test_opts
+    and vim.list_extend(command, test_opts)
+    or command
 end
 
 ---@async